// src/api/generated/uberiket/zod.gen.ts
export const zEgenskapVerdi = z.object({
    type: z.string()
});
// ...
export const zGyldighetsperiode = z.object({
    startdato: z.iso.date(),
    sluttdato: z.optional(z.iso.date())
});
// ...
export const zSideMetadata = z.object({
    returnert: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    sidestorrelse: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    neste: z.optional(zNesteSide)
});

export const zSideposisjon = z.enum([
    'V',
    'H',
    'HV',
    'M',
    'K',
    'MV',
    'MH',
    'VT',
    'HT',
    'R0',
    'L'
]);

export const zStedfesting = z.object({
    type: z.string()
});
// ...
export const zVegobjekt = z.object({
    id: z.coerce.bigint().min(BigInt('-9223372036854775808'), { error: 'Invalid value: Expected int64 to be >= -9223372036854775808' }).max(BigInt('9223372036854775807'), { error: 'Invalid value: Expected int64 to be <= 9223372036854775807' }),
    versjon: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    typeId: z.int().min(-2147483648, { error: 'Invalid value: Expected int32 to be >= -2147483648' }).max(2147483647, { error: 'Invalid value: Expected int32 to be <= 2147483647' }),
    gyldighetsperiode: z.optional(zGyldighetsperiode),
    egenskaper: z.optional(z.record(z.string(), zEgenskapVerdi)),
    barn: z.optional(z.record(z.string(), z.array(z.coerce.bigint().min(BigInt('-9223372036854775808'), { error: 'Invalid value: Expected int64 to be >= -9223372036854775808' }).max(BigInt('9223372036854775807'), { error: 'Invalid value: Expected int64 to be <= 9223372036854775807' })))),
    stedfesting: z.optional(zStedfesting),
    sistEndret: z.iso.datetime()
});
// ...
export const zVeglenkesekvens = z.object({
    id: z.coerce.bigint().min(BigInt('-9223372036854775808'), { error: 'Invalid value: Expected int64 to be >= -9223372036854775808' }).max(BigInt('9223372036854775807'), { error: 'Invalid value: Expected int64 to be <= 9223372036854775807' }),
    porter: z.optional(z.array(zVeglenkesekvensport)),
    veglenker: z.optional(z.array(zVeglenke)),
    lengde: z.optional(z.number()),
    sistEndret: z.optional(z.iso.datetime())
});

export const zVeglenkesekvenserSide = z.object({
    veglenkesekvenser: z.array(zVeglenkesekvens),
    metadata: zSideMetadata
});
// ...
export const zHentVeglenkesekvenserResponse = zVeglenkesekvenserSide;